feat: Move Database to its own module
